qasyncserial/QAsyncSerial.cpp: Warning gefixt
SerialPort.h: Neue Funktion die den korrekten Gerätenamen zurück gibt (bisher nur Windows)
SerialPortHandler: Bekommt nun SerialPort im Konstruktor
helper_functions.h/cpp: Neue Hilfsfunktion die einen int in einzelne Bytes konvertieren kann (als QByteArray)
messages/pingmessage: Typ der sequenceNumber zu einem sinnigeren uint16_t geändert, sequenceNumber nicht mehr als ASCII sondern als QByteArray ausgeben
messages/basemessage: Benutzt nun auch intToQByteArray
mainwindow: Button zum Senden einer Ping-Nachricht

+++ messages/pingmessage.h
@@ -1,6 +1,7 @@
 #ifndef PINGMESSAGE_H
 #define PINGMESSAGE_H
 
+#include <stdint.h>
 #include "basemessage.h"
 
 class PingMessage : public BaseMessage
@@ -8,10 +9,10 @@ class PingMessage : public BaseMessage
 public:
     PingMessage();
 
-	int sequenceNumber();
+	uint16_t sequenceNumber();
 
 protected:
-	int _sequenceNumber;
+	uint16_t _sequenceNumber;
 
 private:
 	QByteArray prepareData();
